refactor(#34): invert control for encryption entities
